third_party/WebKit/LayoutTests/fast/dom/shadow/get-distributed-nodes-orphan.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>

<div id="container"></div>
<pre id="console"></pre>

<script>
description("getDistributedNodes() shouldn't work unless it is inDocument");

function assertNodeList(nodeList, expectedNodes)
{
    window.nodeList = nodeList;
    window.expectedNodes = expectedNodes;
    shouldBe("nodeList.length", "expectedNodes.length");
    for (var i = 0; i < nodeList.length; ++i) {
        shouldBe("nodeList.item(" + i + ")", "expectedNodes[" + i + "]");
    }
}

var host = document.createElement('div');
var shadowRoot = host.createShadowRoot();
var child = document.createElement('div');
var rootChild = document.createElement('div');
var content = document.createElement('content');

host.appendChild(child);
rootChild.appendChild(content);
shadowRoot.appendChild(rootChild);

assertNodeList(content.getDistributedNodes(), []);
debug('');
</script>
</body>
</html>
