third_party/WebKit/LayoutTests/fast/dom/shadow/get-distributed-nodes-orphan-expected.txt

getDistributedNodes() shouldn't work unless it is inDocument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length

PASS successfullyParsed is true

TEST COMPLETE
